detector/requirements.txt
chardet==4.0.0
Django==4.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.3
opencv-contrib-python==4.7.0.72
opencv-python-headless==4.7.0.72
pandas==1.4.2
Pillow==9.4.0
tqdm==4.64.0
tensorflow==2.12.0